requirements.txt
mlx-whisper>=0.4.0
sounddevice>=0.5.0
numpy>=1.24.0
pyperclip>=1.9.0
ollama>=0.4.0
pyobjc-framework-Quartz>=10.0
